+++ src/app/app.module.ts
@@ -15,6 +15,10 @@ import { environment } from '../environments/environment';
 import { TheOrangeAllianceComponent } from './app.component';
 import { HomeComponent } from './views/home/home.component';
 import { AddDataComponent } from './views/add-data/add-data.component';
+import { ManageEventsComponent } from './views/account/subpages/manage-events/manage-events.component';
+import { UsersComponent } from './views/account/subpages/users/users.component';
+import { CreateEventComponent } from './views/account/subpages/create-event/create-event.component';
+import { CacheComponent } from './views/account/subpages/cache/cache.component';
 import { RetrieverComponent } from './views/account/subpages/retriever/retriever.component';
 import { AccountComponent } from './views/account/account.component';
 import { EventsComponent } from './views/events/events.component';
@@ -61,6 +65,10 @@ export function HttpLoaderFactory(httpClient: HttpClient) {
     TheOrangeAllianceComponent,
     HomeComponent,
     AddDataComponent,
+    ManageEventsComponent,
+    UsersComponent,
+    CreateEventComponent,
+    CacheComponent,
     RetrieverComponent,
     AccountComponent,
     LoginComponent,
