+++ src/app/providers/cloud-functions.ts
@@ -10,7 +10,7 @@ export class CloudFunctions {
   // private baseUrl = 'http://localhost:5000/the-orange-alliance/us-central1/requireValidations'; // Tests Only
 
   constructor(private http: HttpClient) {
-    
+
   }
 
   public getUserData(user: User): Promise<TOAUser> {
